223 Replaces the old Timeline with the custom made one
